Implement the ls command (#20)
# Description

This PR adds support for listing GPG keys in the user's keyring. The
command has the following options: [-a, -p, -n] to customize the keys
that will be returned.

Only the most common key configurations are currently supported. More
support will be added as needed, based on user feedback.

Also, testing against python 3.9 has been disabled for now because of a
bug in pylint.

Release: #minor